прикрутил слайсы в проект
