+++ src/pages/offer-screen/offer-screen.tsx
@@ -1,7 +1,7 @@
 import { useParams } from 'react-router-dom';
 import ReviewForm from '../../components/review-form/review-form.tsx';
 import BookmarkButton from '../../components/bookmark-button/bookmark-button.tsx';
-import { AuthorizationStatus, BookmarkButtonDisplayMode, CardDisplayMode, NameSpace } from '../../const.ts';
+import { AuthorizationStatus, BookmarkButtonDisplayMode, CardDisplayMode } from '../../const.ts';
 import ReviewsList from '../../components/reviews-list/reviews-list.tsx';
 import PlaceList from '../../components/place-list/place-list.tsx';
 import Map from '../../components/map/map.tsx';
@@ -10,6 +10,7 @@ import NotFoundScreen from '../not-found-screen/not-found-screen.tsx';
 import { fetchComments, fetchNearOffersAction, getOfferAction } from '../../store/api-actions.ts';
 import { useEffect } from 'react';
 import { getAuthorizationStatus } from '../../store/user-process/selectors.ts';
+import { NameSpace } from '../../store/const.ts';
 
 function OfferScreen(): JSX.Element {
   const {id} = useParams();
