+++ src/store/const.ts
@@ -1,0 +1,14 @@
+export enum APIRoute {
+  Offers = '/offers',
+  Login = '/login',
+  Logout = '/logout',
+  NearOffers = '/offers/{offerId}/nearby',
+  Comments = '/comments/',
+}
+
+export enum NameSpace {
+  User = 'USER',
+  Data = 'DATA',
+  City = 'CITY',
+  Sort = 'SORT',
+}
