+++ src/store/root-reducer.ts
@@ -1,9 +1,9 @@
-import { combineReducers } from "@reduxjs/toolkit";
-import { NameSpace } from "../const";
-import { city } from "./city/city";
-import { dataProcess } from "./data-process/data-process";
-import { sort } from "./sort/sort";
-import { userProcess } from "./user-process/user-process";
+import { combineReducers } from '@reduxjs/toolkit';
+import { city } from './city/city';
+import { dataProcess } from './data-process/data-process';
+import { sort } from './sort/sort';
+import { userProcess } from './user-process/user-process';
+import { NameSpace } from './const';
 
 export const rootReducer = combineReducers({
   [NameSpace.City]: city.reducer,
