+++ src/store/user-process/selectors.ts
@@ -1,5 +1,6 @@
-import { AuthorizationStatus, NameSpace } from "../../const";
-import { State } from "../../types/state";
+import { AuthorizationStatus } from '../../const';
+import { State } from '../../types/state';
+import { NameSpace } from '../const';
 
 export const getAuthorizationStatus = (state: State): AuthorizationStatus => state[NameSpace.User].authorizationStatus;
 export const getAuthCheckedStatus = (state: State): boolean => state[NameSpace.User].authorizationStatus !== AuthorizationStatus.Unknown;
